Completed calculateDifference() in TrianglePictureFilter
Completed the method so the class can get the percentage of similarity between the original and the would-be copy, if the primitive was to be added.
Please note that the typecast will not cause an overflow, as the value can be only between 0 and 100. Furthermore, I chose a long integer, as png files do not really have a maximum resolution, so all available pixels might have caused an overflow with just a regular integer.